Apply background color to tray windows
Transparency of tray windows was not working correctly. The tray windows
would seem to take on a random background color.

I tried a lot of variations to try to get working transparency, but did
not have much luck.

The following do work, though they aren't true transparency:
- You can set the bg color of the tray window itself.
- You can create a "frame window" for each tray icon, and give that
  frame a background color. For some reason, the child tray windows
  seem to have correct transparency over such a parent.

There is certainly room for improvement here, but fixing the bg color
to match the React tray bg is an improvement from the prior state.